pyproject.toml
[tool.poetry]
name = "anser_now"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.0"
sqlmodel = "^0.0.18"
injector = "^0.21.0"
pydantic = "^2.7.1"
alembic = "^1.13.1"
aiomysql = "^0.2.0"
python-dotenv = "^1.0.1"
uvicorn = "^0.29.0"
authlib = "^1.3.0"
passlib = "^1.7.4"
websockets = "^12.0"
jinja2 = "^3.1.4"
pytz = "^2024.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"